+++ source/fixes/ChameleonWidescreenFix.cpp
@@ -12,8 +12,8 @@
 using namespace std;
 
 // Constants
-const float kPi = 3.14159265358979323846f;
-const float kTolerance = 0.01f;
+const double kPi = 3.14159265358979323846;
+const double kTolerance = 0.01;
 const streampos kCameraFOVOffset = 0x00098107;
 const streampos kAspectRatioOffset = 0x0009811C;
 
@@ -21,21 +21,21 @@ const streampos kAspectRatioOffset = 0x0009811C;
 int choice1, choice2, tempChoice;
 uint32_t newWidth, newHeight;
 bool fileNotFound, validKeyPressed;
-float currentCameraFOVInRadians, currentCameraFOVInDegrees, newCameraFOVInRadians, newCameraFOVInDegrees, newCameraFOVInDegreesValue, oldWidth = 4.0f, oldHeight = 3.0f, oldHorizontalFOV = 74.999999f, oldAspectRatio = oldWidth / oldHeight, newAspectRatio;
+double currentCameraFOVInRadians, currentCameraFOVInDegrees, newCameraFOVInRadians, newCameraFOVInDegrees, newCameraFOVInDegreesValue, oldWidth = 4.0, oldHeight = 3.0, oldHorizontalFOV = 74.999999, oldAspectRatio = oldWidth / oldHeight, newAspectRatio;
 string descriptor, input;
 fstream file;
 char ch;
 
 // Function to convert degrees to radians
-float DegToRad(float degrees)
+double DegToRad(double degrees)
 {
-    return degrees * (kPi / 180.0f);
+    return degrees * (kPi / 180.0);
 }
 
 // Function to convert radians to degrees
-float RadToDeg(float radians)
+double RadToDeg(double radians)
 {
-    return radians * (180.0f / kPi);
+    return radians * (180.0 / kPi);
 }
 
 // Function to handle user input in choices
@@ -74,7 +74,7 @@ void HandleChoiceInput(int &choice)
     }
 }
 
-void HandleFOVInput(float &newCustomFOVInDegrees)
+void HandleFOVInput(double &newCustomFOVInDegrees)
 {
     do
     {
@@ -84,7 +84,7 @@ void HandleFOVInput(float &newCustomFOVInDegrees)
         // Replaces all commas with dots
         replace(input.begin(), input.end(), ',', '.');
 
-        // Parses the string to a float
+        // Parses the string to a double
         newCustomFOVInDegrees = stof(input);
 
         if (cin.fail())
@@ -157,9 +157,9 @@ void OpenFile(fstream &file, const string &filename)
     }
 }
 
-float NewFOVInDegreesCalculation(uint32_t &newWidthValue, uint32_t &newHeightValue)
+double NewFOVInDegreesCalculation(uint32_t &newWidthValue, uint32_t &newHeightValue)
 {
-    newCameraFOVInDegreesValue = 2.0f * RadToDeg(atan((static_cast<float>(newWidthValue) / static_cast<float>(newHeightValue)) / oldAspectRatio) * tan(DegToRad(oldHorizontalFOV / 2.0f)));
+    newCameraFOVInDegreesValue = 2.0 * RadToDeg(atan((static_cast<double>(newWidthValue) / static_cast<double>(newHeightValue)) / oldAspectRatio) * tan(DegToRad(oldHorizontalFOV / 2.0)));
     return newCameraFOVInDegreesValue;
 }
 
@@ -177,7 +177,7 @@ int main()
         cout << "\n- Enter the desired height: ";
         HandleResolutionInput(newHeight);
 
-        newAspectRatio = static_cast<float>(newWidth) / static_cast<float>(newHeight);
+        newAspectRatio = static_cast<double>(newWidth) / static_cast<double>(newHeight);
 
         cout << "\n- Do you want to set the camera FOV automatically based on the typed resolution above (1) or set a custom camera FOV value (2)?: ";
         HandleChoiceInput(choice1);
