+++ source/fixes/SASAntiTerrorForceFOVChanger.cpp
@@ -19,7 +19,7 @@ fstream file;
 string input;
 int choice, tempChoice;
 bool fileNotFound, validKeyPressed;
-float currentCameraFOV, newCameraFOV;
+double currentCameraFOV, newCameraFOV;
 char ch;
 
 // Function to handle user input in choices
@@ -58,7 +58,7 @@ void HandleChoiceInput(int &choice)
     }
 }
 
-float HandleFOVInput()
+double HandleFOVInput()
 {
     do
     {
@@ -68,7 +68,7 @@ float HandleFOVInput()
         // Replaces all commas with dots
         replace(input.begin(), input.end(), ',', '.');
 
-        // Parses the string to a float
+        // Parses the string to a double
         customFOV = stof(input);
 
         if (cin.fail())
